2 robots,  each on same map,  2nd robot overwrites paths
aStar2 and implementation2  both have two working robots,( made new
files to not break working ones)  the code is not good or pretty! but it
kinda works.  2nd robot overwirtes the first ones path,  and the output
arrows and stuff is a fucking mess, commented out chosing a path for
testing  ,needs work but its progress i suppose