check extended description
Only changed the "NEWaStarSearch" file, 
 Added user input for start and end for both robots.  
To change what map you want to use change the file names location for the traps and the wall, its in the comments(We could just have 4 slightly different aStar files each with a different map in it, just a thought).  
ive created the maps that he wanted us to test,  there in the "GridFiles" folder,  also created a 4th map and trap file.  
The file called "Star and End locations.txt" contains the start and end location for all the maps,  also a wee bit of info about the 4th map i created(when we had it in we should give him this file too so he knows whats going on)